+++ server/app/model/system/Dictionary.php
@@ -35,6 +35,7 @@ class Dictionary extends Model
      */
     public function getStatusTextAttr($value, $data): string
     {
+        if (!isset($data['status'])) return '';
         return $this->getStatusText($data['status'], [1 => '正常', 0 => '禁用']);
     }
 }